requirements.txt
alabaster==0.7.13
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
docutils==0.18.1
ghp-import==2.1.0
idna==3.4
ImageHash==4.3.1
imagesize==1.4.1
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.0
packaging==23.1
pandas==2.1.0
pdqhash==0.2.5
Pillow==10.0.1
Pygments==2.16.1
requests==2.31.0
snowballstemmer==2.2.0
scikit-learn==1.3.0
scipy==1.11.2
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
tqdm==4.66.1   
urllib3==2.0.4
